pyproject.toml
[project]
name = "integracao-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.2",
    "pandas>=2.3.3",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.21.1",
]
